pyproject.toml
[project]
name = "ontology-agent"
version = "0.1.0"
description = "FIBO knowledge extraction MVP built with the Gemini ADK"
authors = [{ name = "Seocho Graph-Ontology Agent Team" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "google-adk>=1.8.0",
    "python-dotenv>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.sdist]
include = ["app", "prompts", "scripts", "samples", "MVP_GUIDE.md", "README.md", "pyproject.toml"]

[tool.ruff]
line-length = 100
target-version = "py311"
